pyproject.toml
[project]
name = "calendar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.8",
    "google-api-python-client>=2.167.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "openai>=1.75.0",
    "openai-agents>=0.0.11",
    "pinecone>=6.0.2",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
]
